move correlateRC functions into a helper file
